% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{any.duplicated}
\alias{any.duplicated}
\title{any.duplicated}
\usage{
\method{any}{duplicated}(vec, summarize = TRUE, max.shown = 25)
}
\arguments{
\item{vec}{input vector}

\item{summarize}{Print summary? Default: TRUE}

\item{max.shown}{Print first X elements. Default: 25}
}
\description{
How many entries are duplicated?.
}
