% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{arr.of.lists.2.df}
\alias{arr.of.lists.2.df}
\title{arr.of.lists.2.df}
\usage{
arr.of.lists.2.df(two.dim.arr.of.lists)
}
\arguments{
\item{two.dim.arr.of.lists}{PARAM_DESCRIPTION}
}
\description{
Simplify 2D-list-array to a DF.
}
