% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.character.wNames}
\alias{as.character.wNames}
\title{as.character.wNames}
\usage{
\method{as.character}{wNames}(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
Converts your input vector into a character vector, and puts the original character values into the names of the new vector, unless it already has names.
}
