% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.list.df.by.col}
\alias{as.list.df.by.col}
\title{as.list.df.by.col}
\usage{
\method{as.list}{df.by.col}(dtf, na.omit = TRUE, zero.omit = FALSE, omit.empty = FALSE)
}
\arguments{
\item{dtf}{A dataframe.}

\item{na.omit}{Whether to omit rows with missing values.}

\item{zero.omit}{Whether to omit rows with all-zero values.}

\item{omit.empty}{Whether to omit rows with zero length.}
}
\value{
A list.
}
\description{
Split a dataframe into a list by its columns.
}
\examples{
dtf <- data.frame(x = c(1, 2, NA), y = c(3, 4, 0), z = c(5, 6, 7))
as.list.df.by.col(dtf)

}
