% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.listalike}
\alias{as.listalike}
\title{as.listalike}
\usage{
as.listalike(vec, list_wannabe)
}
\arguments{
\item{vec}{PARAM_DESCRIPTION}

\item{list_wannabe}{PARAM_DESCRIPTION}
}
\description{
Convert a vector to a list with certain dimensions, taken from the list it wanna resemble.
}
