% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector.deprecated}
\alias{as.named.vector.deprecated}
\title{as.named.vector.deprecated}
\usage{
as.named.vector.deprecated(df_col, WhichDimNames = 1)
}
\arguments{
\item{df_col}{data frame column}

\item{WhichDimNames}{Shall we extract rows (2) or columns (1, default)?, Default: 1}
}
\description{
Convert a dataframe column or row into a vector, keeping the corresponding dimension name.
}
