% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector.df}
\alias{as.named.vector.df}
\title{as.named.vector.df}
\usage{
as.named.vector.df(df, col.or.row.name.or.index = 1, WhichDimNames = 1, ...)
}
\arguments{
\item{WhichDimNames}{Shall we extract rows (2) or columns (1, default)?, Default: 1}

\item{index.rc}{Which column or row to extract (numeric index).}
}
\description{
Convert any column or row of a dataframe into a vector, keeping the corresponding dimension name.
}
