% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector}
\alias{as.named.vector}
\title{as.named.vector}
\usage{
as.named.vector(df_col, WhichDimNames = 1)
}
\arguments{
\item{df_col}{data frame column}

\item{WhichDimNames}{PARAM_DESCRIPTION, Default: 1}
}
\description{
Convert a dataframe column or row into a vector, keeping the corresponding dimension name.
}
