% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector.table}
\alias{as.named.vector.table}
\title{as.named.vector.table}
\usage{
as.named.vector.table(table, verbose = TRUE, ...)
}
\arguments{
\item{table}{A 1D table.}

\item{verbose}{Print the name of the table that is being used}

\item{...}{Additional arguments passed to \code{as.vector()}.}
}
\description{
Convert a 1D table into a named vector.
}
