% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector2}
\alias{as.named.vector2}
\title{as.named.vector2}
\usage{
as.named.vector2(df, index.rc = 1, WhichDimNames = 1)
}
\arguments{
\item{index.rc}{Which column or row to extract (numeric index).}

\item{WhichDimNames}{Shall we extract rows (2) or columns (1, default)?, Default: 1}
}
\description{
Convert any column or row of a dataframe into a vector, keeping the corresponding dimension name.
}
