% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.numeric.wNames.character}
\alias{as.numeric.wNames.character}
\title{as.numeric.wNames.character}
\usage{
\method{as.numeric}{wNames.character}(vec, verbose = TRUE, factor.to.character = TRUE, ...)
}
\arguments{
\item{vec}{input vector}

\item{verbose}{Print troubleshooting messages}

\item{factor.to.character}{convert Input vector to first to 'character', then numeric.}

\item{...}{Pass any other argument to as.numeric()}
}
\description{
Converts (1) a 'character' v. into a numeric v., or
a 'factor' v. via as.numeric(as.character(vec)) and preserves the original names.
The old 'as.numeric.wNames()' is deprecated as it was not clearly documented that it converts via
factor in any case. Code saved at the end.
}
\examples{
vec <- as.character(c(1, 2, 8, 9))
names(vec) <- LETTERS[1:4]
vec
as.numeric.wNames.character(vec)
vec2 <- as.factor(c(1, 2, 8, 9))
names(vec2) <- LETTERS[1:4]
vec2
as.numeric.wNames.character(vec2, factor.to.character = FALSE)

}
