% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.numeric.wNames.factor}
\alias{as.numeric.wNames.factor}
\title{as.numeric.wNames.factor}
\usage{
\method{as.numeric}{wNames.factor}(vec, ...)
}
\arguments{
\item{vec}{vector of factors, strings, (or even logical)}

\item{...}{Pass any other argument to as.factor()}
}
\description{
Turn any vector into numeric categories as.numeric(as.factor(vec))
Forerly as.factor.numeric
}
\examples{
as.numeric.wNames.factor(LETTERS[1:4])

}
