% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.numeric.wNames}
\alias{as.numeric.wNames}
\title{as.numeric.wNames}
\usage{
\method{as.numeric}{wNames}(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
Converts any vector into a numeric vector, and puts the original character values into the names of the new vector, unless it already has names. Useful for coloring a plot by categories, name-tags, etc.
}
