% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{checkMinOverlap}
\alias{checkMinOverlap}
\title{Check Minimum Overlap Between Two Vectors}
\usage{
checkMinOverlap(x, y, min_overlap = 0.2, stop_it = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{First character vector.}

\item{y}{Second character vector.}

\item{min_overlap}{Minimum required overlap as a fraction of the shorter vector's length.}

\item{stop_it}{Use stop(), else warning()}

\item{verbose}{verbose}
}
\value{
Invisible TRUE if the overlap is sufficient, otherwise stops with an error.
}
\description{
Checks if the overlap between two character vectors is at least a specified
percentage of the shorter vector. Stops execution with an error if the condition is not met.
}
