% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{clip.outliers}
\alias{clip.outliers}
\title{clip.outliers}
\usage{
clip.outliers(
  valz,
  high = TRUE,
  probs = c(0.01, 0.99),
  na.rm = TRUE,
  showhist = FALSE,
  ...
)
}
\arguments{
\item{valz}{PARAM_DESCRIPTION}

\item{high}{PARAM_DESCRIPTION, Default: TRUE}

\item{probs}{PARAM_DESCRIPTION, Default: c(0.01, 0.99)}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{showhist}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\description{
Signal clipping based on the input data's distribution. It clips values above or below the extreme N\% of the distribution.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
