% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{clip.values}
\alias{clip.values}
\title{clip.values}
\usage{
clip.values(valz, high = TRUE, thr = 3)
}
\arguments{
\item{valz}{PARAM_DESCRIPTION}

\item{high}{PARAM_DESCRIPTION, Default: TRUE}

\item{thr}{PARAM_DESCRIPTION, Default: 3}
}
\description{
Signal clipping. Cut values above or below a threshold.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
