% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{col2named.vec.tbl}
\alias{col2named.vec.tbl}
\title{col2named.vec.tbl}
\usage{
col2named.vec.tbl(tbl.2col)
}
\arguments{
\item{tbl.2col}{PARAM_DESCRIPTION}
}
\description{
Convert a 2-column table(data frame) into a named vector. 1st column will be used as names.
}
