% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{col2named.vector}
\alias{col2named.vector}
\title{col2named.vector}
\usage{
col2named.vector(df_col)
}
\arguments{
\item{df_col}{data frame column}
}
\description{
Convert a dataframe column into a vector, keeping the corresponding dimension name.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
