% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colCV}
\alias{colCV}
\title{colCV}
\usage{
colCV(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input matrix, or all-numeric dataframe.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the CV of each column of a numeric matrix / data frame.
}
