% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colIQR}
\alias{colIQR}
\title{colIQR}
\usage{
colIQR(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the SEM of each column of a numeric matrix / data frame.
}
