% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colMax}
\alias{colMax}
\title{colMax}
\usage{
colMax(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the maximum of each column of a numeric matrix / data frame.
}
