% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colMin}
\alias{colMin}
\title{colMin}
\usage{
colMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the minimum of each column of a numeric matrix / data frame.
}
