% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colMultiply}
\alias{colMultiply}
\title{Column-wise multiplication of a matrix by a vector}
\usage{
colMultiply(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix.}

\item{vec}{Vector to multiply by.}
}
\description{
Multiply each column of a matrix by the corresponding element of a vector.
}
