% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colMutliply}
\alias{colMutliply}
\title{colMutliply}
\usage{
colMutliply(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to multiply by.}
}
\description{
See more: https://stackoverflow.com/questions/20596433/how-to-divide-each-row-of-a-matrix-by-elements-of-a-vector-in-r.
}
