% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colNameMatrix}
\alias{colNameMatrix}
\title{colNameMatrix}
\usage{
colNameMatrix(mat_w_dimnames)
}
\arguments{
\item{mat_w_dimnames}{A named matrix to copy from.}
}
\description{
Create a copy of your matrix, where every entry is replaced by the corresponding column name. Useful if you want to color by column name in a plot (where you have different number of NA-values in each column).
}
