% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colquantile}
\alias{colquantile}
\title{colquantile}
\usage{
colquantile(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\description{
Calculates the SEM of each column of a numeric matrix / data frame.
}
