% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colSD}
\alias{colSD}
\title{colSD}
\usage{
colSD(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the SD of each column of a numeric matrix / data frame.
}
