% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colsplit}
\alias{colsplit}
\title{colsplit}
\usage{
colsplit(df, f = colnames(df))
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION, Default: colnames(df)}
}
\description{
Split a data frame by a factor corresponding to columns.
}
