% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colSubtract}
\alias{colSubtract}
\title{colSubtract}
\usage{
colSubtract(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix.}

\item{vec}{Vector to subtract. Length = nr. columns.}
}
\description{
Subtract a vector (length = nr. columns) column by column from each value of the matrix.
}
