% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{combine.matrices.by.rowname.intersect}
\alias{combine.matrices.by.rowname.intersect}
\title{combine.matrices.by.rowname.intersect}
\usage{
combine.matrices.by.rowname.intersect(matrix1, matrix2, k = 2)
}
\arguments{
\item{matrix1}{A matrix.}

\item{matrix2}{A matrix.}

\item{k}{The number of rows to print from the matrices with the most missing values.}
}
\value{
A matrix with the rows of \code{matrix1} and \code{matrix2} that intersect.
}
\description{
Combine two matrices by the intersection of their row names.
}
