% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{copy.dimension.and.dimnames}
\alias{copy.dimension.and.dimnames}
\title{copy.dimension.and.dimnames}
\usage{
copy.dimension.and.dimnames(list.1D, obj.2D)
}
\arguments{
\item{list.1D}{A 1D vector.}

\item{obj.2D}{A 2D array.}
}
\value{
A 2D array with the same dimension and dimnames as \code{obj.2D}.
}
\description{
Copy the dimension and dimnames of a 1D vector to a 2D array.
}
