% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{count_occurrence_each_element}
\alias{count_occurrence_each_element}
\title{count_occurrence_each_element}
\usage{
count_occurrence_each_element(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
Count the number of times each element occurs in the full vector, AND give it back as a vector, that is the same length as the input vector, each element corresponding one-by-one.
}
