% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{cumsubtract}
\alias{cumsubtract}
\title{cumsubtract}
\usage{
cumsubtract(numericVec = blanks)
}
\arguments{
\item{numericVec}{PARAM_DESCRIPTION, Default: blanks}
}
\description{
Cumulative subtraction, opposite of cumsum().
}
