% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{df.col.2.named.vector}
\alias{df.col.2.named.vector}
\title{df.col.2.named.vector}
\usage{
df.col.2.named.vector(df, col, names = NULL)
}
\arguments{
\item{df}{data frame or tibble}

\item{col}{column index or name for values.}

\item{names}{optional column index for names, Default: NULL (then rownames are used)}
}
\description{
Convert a dataframe column into a vector, keeping the corresponding dimension name.
}
