% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{df.row.2.named.vector}
\alias{df.row.2.named.vector}
\title{df.row.2.named.vector}
\usage{
df.row.2.named.vector(df, row, names = NULL)
}
\arguments{
\item{df}{data frame or tibble}

\item{row}{row index or name for values.}

\item{names}{optional column index for names, Default: NULL (then colnames are used)}
}
\description{
Convert a dataframe row into a vector, keeping the corresponding dimension name.
}
