% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{dput_pretty}
\alias{dput_pretty}
\title{Pretty Printing of R Objects (pretty_dput)}
\usage{
dput_pretty(vec)
}
\arguments{
\item{vec}{A named vector to be printed. The vector should be in the form \code{c('name' = 'value')}.}
}
\value{
No return value. Outputs the vector with each element on a new line.
}
\description{
This function modifies the output of the traditional dput() function
so that each key-value pair in a vector appears on a new line. It's useful for
creating more readable output.
}
\examples{
vec <- c(`0` = "ACyte", `1` = "Misp.1.DCN")
dput_pretty(vec)

}
