% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{fano}
\alias{fano}
\title{fano}
\usage{
fano(x, na.rm = TRUE, USE = "na.or.complete")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{USE}{PARAM_DESCRIPTION, Default: 'na.or.complete'}
}
\description{
Calculates the fano factor on a numeric vector (it excludes NA-s by default).
}
