% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{filter_survival_length}
\alias{filter_survival_length}
\title{filter_survival_length}
\usage{
filter_survival_length(length_new, length_old, prepend = "")
}
\arguments{
\item{length_new}{PARAM_DESCRIPTION}

\item{length_old}{PARAM_DESCRIPTION}

\item{prepend}{PARAM_DESCRIPTION, Default: ''}
}
\description{
Parse a sentence reporting the \% of filter survival.
}
