% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{fix_tibble_lists}
\alias{fix_tibble_lists}
\title{Convert List Columns of a Tibble to String Vectors}
\usage{
fix_tibble_lists(df, verbose = T, print_full = F, collapse_by = ",")
}
\arguments{
\item{df}{A tibble or data frame where list columns are to be converted to string vectors.}

\item{verbose}{Print anything? Default: \code{TRUE}.}

\item{print_full}{Print full details? Default: \code{FALSE}.}
}
\value{
A tibble or data frame with list columns converted to string vectors.
}
\description{
This function identifies columns of type \code{list} in a tibble or data frame
and converts them to string vectors.
}
\examples{
df <- tibble::tibble(a = list(1:2, 3:4, 5:6), b = c("A", "B", "C"))
fix_tibble_lists(df)

}
