% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{flip_value2name}
\alias{flip_value2name}
\title{flip_value2name}
\usage{
flip_value2name(namedVector, NumericNames = FALSE, silent = F)
}
\arguments{
\item{namedVector}{PARAM_DESCRIPTION}

\item{NumericNames}{PARAM_DESCRIPTION, Default: FALSE}

\item{silent}{PARAM_DESCRIPTION, Default: F}
}
\description{
Flip the values and the names of a vector with names.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
