% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{geomean}
\alias{geomean}
\title{geomean}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the geometric mean of a numeric vector (it excludes NA-s by default).
}
