% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{get.oddoreven}
\alias{get.oddoreven}
\title{get.oddoreven}
\usage{
get.oddoreven(df_ = NULL, rows = FALSE, odd = TRUE)
}
\arguments{
\item{df_}{PARAM_DESCRIPTION, Default: NULL}

\item{rows}{PARAM_DESCRIPTION, Default: FALSE}

\item{odd}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Get odd or even columns or rows of a data frame.
}
