% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{get_col_types}
\alias{get_col_types}
\title{Extract and Display Column Types of a Data Frame or Tibble}
\usage{
get_col_types(df, print_it = TRUE)
}
\arguments{
\item{df}{A data frame or tibble whose column types are to be extracted.}

\item{print_it}{Logical; if \code{TRUE} (default), prints a table of column types and a summary.}
}
\value{
A named character vector where names are column names and the values are their respective primary types.
}
\description{
This function returns the primary class/type of each column in a data frame or tibble.
Additionally, it can print a summary of the column types.
}
\examples{
df <- data.frame(a = 1:3, b = c("A", "B", "C"), c = factor(c("X", "Y", "X")))
get_col_types(df)

}
