% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{getCols}
\alias{getCols}
\title{getCols}
\usage{
getCols(
  mat,
  colnamez,
  silent = FALSE,
  removeNAonly = FALSE,
  remove0only = FALSE
)
}
\arguments{
\item{mat}{Input matrix.}

\item{colnamez}{Vector of column names to search for in the matrix.}

\item{silent}{Logical indicating whether to suppress printing of missing columns. Default: FALSE}

\item{removeNAonly}{Logical indicating whether to remove columns with only NA values. Default: FALSE}

\item{remove0only}{Logical indicating whether to remove columns with only zero values. Default: FALSE}
}
\value{
A matrix that is a subset of the input matrix.
}
\description{
Returns a subset of columns based on their names and optionally removes columns with only NA or zero values. Reports the number of missing columns.
}
