% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{getProject}
\alias{getProject}
\title{getProject}
\usage{
getProject()
}
\value{
The final subfolder of your project, or \code{NULL} if you are not running one
}
\description{
Try to get the project name you are working on in RStudio.
}
\examples{
getProject()

}
