% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{getRows}
\alias{getRows}
\title{getRows}
\usage{
getRows(
  mat,
  rownamez,
  silent = FALSE,
  removeNAonly = FALSE,
  remove0only = FALSE
)
}
\arguments{
\item{mat}{Input matrix.}

\item{rownamez}{Vector of row names to search for in the matrix.}

\item{silent}{Logical indicating whether to suppress printing of missing rows. Default: FALSE}

\item{removeNAonly}{Logical indicating whether to remove rows with only NA values. Default: FALSE}

\item{remove0only}{Logical indicating whether to remove rows with only zero values. Default: FALSE}
}
\value{
A matrix that is a subset of the input matrix.
}
\description{
Returns a subset of rows based on their names and optionally removes rows with only NA or zero values. Reports the number of missing rows.
}
