% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{grepv}
\alias{grepv}
\title{grepv}
\usage{
grepv(
  pattern,
  x,
  ignore.case = FALSE,
  perl = FALSE,
  value = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  invert = FALSE,
  ...
)
}
\arguments{
\item{pattern}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{ignore.case}{PARAM_DESCRIPTION, Default: FALSE}

\item{perl}{PARAM_DESCRIPTION, Default: FALSE}

\item{value}{PARAM_DESCRIPTION, Default: FALSE}

\item{fixed}{PARAM_DESCRIPTION, Default: FALSE}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}

\item{invert}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\description{
grep returning the value.
}
