% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{idim}
\alias{idim}
\title{idim}
\usage{
idim(any_object)
}
\arguments{
\item{any_object}{PARAM_DESCRIPTION}
}
\description{
A dim() function that can handle if you pass on a vector: then, it gives the length.
}
\examples{
\dontrun{
if(interactive()){
 # a <- 1:2; idim(a)
 }
}
}
