% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intermingle2lists}
\alias{intermingle2lists}
\title{intermingle2lists}
\usage{
intermingle2lists(L1, L2)
}
\arguments{
\item{L1}{PARAM_DESCRIPTION}

\item{L2}{PARAM_DESCRIPTION}
}
\description{
Combine 2 lists (of the same length) so that form every odd and every even element of a unified list. Useful for side-by-side comparisons, e.g. in wstripchart_list().
}
