% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intermingle2vec}
\alias{intermingle2vec}
\title{intermingle2vec}
\usage{
intermingle2vec(V1, V2, wNames = TRUE)
}
\arguments{
\item{V1}{PARAM_DESCRIPTION}

\item{V2}{PARAM_DESCRIPTION}

\item{wNames}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Combine 2 vectors (of the same length) so that form every odd and every even element of a unified vector.
}
