% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intersect.ls}
\alias{intersect.ls}
\title{intersect.ls}
\usage{
intersect.ls(ls, ...)
}
\arguments{
\item{ls}{List of 2 or more vectors (sets) with categorical variables.}

\item{...}{Pass any other argument.}
}
\description{
Intersect any number of lists.
}
