% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{is.list.simple}
\alias{is.list.simple}
\title{Test if object is a list}
\usage{
is.list.simple(object, warn_nested = TRUE)
}
\arguments{
\item{object}{Object to test.}

\item{warn_nested}{Warn if nested lists are detected? Default: TRUE}
}
\description{
The 'is.list()' function "fails: on tibbles and data frames: it returns TRUE,
as if it were a list. This distinguishes and identifies simple lists. That's why we need this function.
}
\examples{
is.list.simple(list())
is.list.simple(dplyr::tibble())
}
