% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{iterBy.over}
\alias{iterBy.over}
\title{iterBy.over}
\usage{
iterBy.over(vec, by = 9)
}
\arguments{
\item{vec}{Input vector}

\item{by}{PARAM_DESCRIPTION, Default: 9}
}
\description{
Iterate over a vector by every N-th element.
}
