% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.from.list}
\alias{list.from.list}
\title{list.from.list}
\usage{
list.from.list(template, fill = NaN)
}
\arguments{
\item{template}{A template list with names,.}

\item{fill}{The value to fill the new vector, Default: NaN}
}
\description{
create an empty list from named and structured as the template list.
}
