% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.from.template}
\alias{list.from.template}
\title{list.from.template}
\usage{
list.from.template(x, fill = NA)
}
\arguments{
\item{x}{A template list with names,.}

\item{fill}{The value to fill the new list, Default: NA}
}
\description{
Create an empty list from a template list, copying names and filling values with NA.
}
