% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.fromNames}
\alias{list.fromNames}
\title{list.fromNames}
\usage{
list.fromNames(name_vec = LETTERS[1:5], fill = NaN)
}
\arguments{
\item{name_vec}{A vector of names, Default: LETTERS\link{1:5}}

\item{fill}{The value to fill the new vector, Default: NaN}
}
\description{
create list from a vector with the names of the elements.
}
