% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.wNames}
\alias{list.wNames}
\title{list.wNames}
\usage{
list.wNames(...)
}
\arguments{
\item{...}{Pass any other argument.}
}
\description{
Create a list with names from ALL variables you pass on to the function.
}
