% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list2fullDF.presence}
\alias{list2fullDF.presence}
\title{list2fullDF.presence}
\usage{
list2fullDF.presence(
  your.list = list(set.1 = LETTERS[1:5], set.2 = LETTERS[3:9]),
  byRow = TRUE,
  FILL = 0
)
}
\arguments{
\item{your.list}{PARAM_DESCRIPTION, Default: list(set.1 = LETTERS\link{1:5}, set.2 = LETTERS\link{3:9})}

\item{byRow}{PARAM_DESCRIPTION, Default: TRUE}

\item{FILL}{PARAM_DESCRIPTION, Default: 0}
}
\description{
Convert a list to a full matrix.  Designed for occurence counting, think tof table(). Rows = all ENTRIES of within your list, columns = names(your_list).
}
