% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{lookup}
\alias{lookup}
\title{lookup}
\usage{
lookup(needle, haystack, exact = TRUE, report = FALSE)
}
\arguments{
\item{needle}{PARAM_DESCRIPTION}

\item{haystack}{PARAM_DESCRIPTION}

\item{exact}{PARAM_DESCRIPTION, Default: TRUE}

\item{report}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
Awesome pattern matching for a set of values in another set of values. Returns a list with all kinds of results.
}
