% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{ls2categvec}
\alias{ls2categvec}
\title{ls2categvec}
\usage{
ls2categvec(your_list)
}
\arguments{
\item{your_list}{A list.}
}
\value{
A vector with the same elements as \code{your_list}, but with the names of the list elements repeated as many times as the number of elements in each list element.
}
\description{
Converts a list to a vector repeating list-element names, while vector names are the list elements.
}
\examples{
ls2categvec(list(a = 1, b = 2, c = 3))
}
