% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{matrix.fromNames}
\alias{matrix.fromNames}
\title{matrix.fromNames}
\usage{
matrix.fromNames(rowname_vec = 1:10, colname_vec = LETTERS[1:5], fill = NA)
}
\arguments{
\item{rowname_vec}{A vector of names for rows, Default: \code{1:10}}

\item{colname_vec}{A vector of names for columns, Default: \code{LETTERS[1:5]}}

\item{fill}{The value to fill the new vector, Default: \code{NA}}
}
\description{
Create a matrix from 2 vectors defining the row- and column names of the matrix. Default fill value: NA.
}
