% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{mdlapply}
\alias{mdlapply}
\title{mdlapply}
\usage{
mdlapply(list_2D, ...)
}
\arguments{
\item{list_2D}{A multidimensional array.}

\item{...}{Function and arguments to pass to \code{lapply()}.}
}
\value{
A multidimensional array with the same dimensions as \code{list_2D}.
}
\description{
A wrapper for \code{lapply()} that works on multidimensional arrays.
}
