% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{mdlapply2df}
\alias{mdlapply2df}
\title{mdlapply2df}
\usage{
mdlapply2df(list_2D, ...)
}
\arguments{
\item{list_2D}{A multidimensional array.}

\item{...}{Pass any other argument.}
}
\description{
Multi dimensional lapply + arr.of.lists.2.df (simplify 2D-list-array to a DF).
}
