% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{mean_normalize}
\alias{mean_normalize}
\title{mean_normalize}
\usage{
mean_normalize(mat)
}
\arguments{
\item{mat}{Numeric input matrix.}
}
\description{
Normalize each column to the mean of the columns.
}
