% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{mean_of_log}
\alias{mean_of_log}
\title{mean_of_log}
\usage{
mean_of_log(x, k = 2, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION, Default: 2}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the mean of the log_k of a numeric vector (it excludes NA-s by default).
}
