% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{median_normalize}
\alias{median_normalize}
\title{median_normalize}
\usage{
median_normalize(mat)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}
}
\description{
Normalize each column to the median of all the column-sums.
}
