% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_1col_dfs_by_rn}
\alias{merge_1col_dfs_by_rn}
\title{merge_1col_dfs_by_rn}
\usage{
merge_1col_dfs_by_rn(list_of_dfs, FILLwith = 0, columnUSE = 1)
}
\arguments{
\item{list_of_dfs}{list of 1col dfs}

\item{FILLwith}{0 by def}

\item{columnUSE}{column index in both. 1 by default.}
}
\description{
A function that merges single-column data frames based on rownames.
}
\examples{
df1 <- data.frame(val = c(1, 2, 3), row.names = c("a", "b", "c"))
df2 <- data.frame(val = c(4, 5), row.names = c("b", "d"))
merge_1col_dfs_by_rn(list(A = df1, B = df2))
}
