% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_2_named_vec_as_df}
\alias{merge_2_named_vec_as_df}
\title{merge_2_named_vec_as_df}
\usage{
merge_2_named_vec_as_df(x, y, as.tibble = FALSE)
}
\arguments{
\item{x}{A vector with named elements.}

\item{y}{Another vector with named elements.}

\item{as.tibble}{Logical; if TRUE, returns a tibble instead of a data frame. Default: FALSE}
}
\description{
Merge two named vectors by names, into a dataframe with 2 columns.
}
\examples{
# a <- 1:5; names(a) <- letters[a]; b <- 9:3; names(b) <- letters[b]; merge_2_named_vec_as_df(a,b)
}
