% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_ls_of_named_vec_as_df_cols}
\alias{merge_ls_of_named_vec_as_df_cols}
\title{merge_ls_of_named_vec_as_df_cols}
\usage{
merge_ls_of_named_vec_as_df_cols(
  named_list = list(vec1 = c(A = 1, B = 2, C = 3), vec2 = c(B = 4, D = 5), vec3 = c(A =
    6, C = 7, D = 8), vec4 = c(B = 9, C = 10, D = 11, E = 12)),
  missing_values = NaN
)
}
\arguments{
\item{named_list}{A named list of named vectors.}

\item{missing_values}{How to fill missing values}
}
\description{
Merge any number of named vectors (presented as a list) by names, into a dataframe
}
\examples{
# merge_ls_of_named_vec_as_df_cols()
}
