% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_numeric_df_by_rn}
\alias{merge_numeric_df_by_rn}
\title{merge_numeric_df_by_rn}
\usage{
merge_numeric_df_by_rn(x, y)
}
\arguments{
\item{x}{Input matrix, or all-numeric dataframe.}

\item{y}{Input matrix, or all-numeric dataframe.}
}
\description{
Merge 2 numeric data frames by rownames.
}
