% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{most_frequent_elements}
\alias{most_frequent_elements}
\title{most_frequent_elements}
\usage{
most_frequent_elements(vec, topN = 10)
}
\arguments{
\item{vec}{input vector}

\item{topN}{How many pof the most frequent elements should be returned? Default: 10}
}
\description{
Show the most frequent elements of a table.
}
