% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{na.omit.strip}
\alias{na.omit.strip}
\title{na.omit.strip}
\usage{
\method{na.omit}{strip}(vec, silent = FALSE)
}
\arguments{
\item{vec}{input vector}

\item{silent}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
Calls na.omit() and returns a clean vector.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{na.omit}}
}
