% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{numerate}
\alias{numerate}
\title{Formats a sequence of numbers with zeropadding.}
\usage{
numerate(
  x = 1,
  y = 100,
  zeropadding = TRUE,
  pad_length = floor(log10(max(abs(x), abs(y)))) + 1
)
}
\arguments{
\item{x}{The starting number.}

\item{y}{The ending number.}

\item{zeropadding}{Whether to use zeropadding (default: TRUE).}

\item{pad_length}{The length of the padded numbers (default: floor(log10(max(abs(x), abs(y)))) + 1).}
}
\value{
A vector of formatted numbers.
}
\description{
Formats a sequence of numbers with zeropadding.
}
\examples{
numerate(1, 122)
}
\seealso{
\code{\link[stringr]{str_pad}}
}
