% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pc_in_total_of_match}
\alias{pc_in_total_of_match}
\title{pc_in_total_of_match}
\usage{
pc_in_total_of_match(vec_or_table, category, NA_omit = TRUE)
}
\arguments{
\item{vec_or_table}{A vector or table.}

\item{category}{The value to calculate the percentage for.}

\item{NA_omit}{Logical. Whether to omit missing values from the calculation.}
}
\value{
The percentage of \code{category} in \code{vec_or_table}.
}
\description{
Calculates the percentage of a certain value within a vector or table.
}
