% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pc_overlap}
\alias{pc_overlap}
\title{Calculate Percentage Overlap Between Two Vectors}
\usage{
pc_overlap(x, y, basis = "x", prefix = NULL, suffix = NULL, ...)
}
\arguments{
\item{x}{The first vector for overlap calculation.}

\item{y}{The second vector for overlap calculation.}

\item{basis}{A character string specifying the basis for calculating the percentage overlap.
Can be "x" for the length of \code{x}, "y" for the length of \code{y}, or "sum" for the sum
of lengths of both \code{x} and \code{y}. Default is "x".}
}
\value{
The percentage of overlap between \code{x} and \code{y} based on the specified basis.
}
\description{
Computes the percentage of overlap between two vectors based on the specified basis of calculation.
}
\examples{
x= 1:5; y=3:8;  pc_overlap(x, y, basis = "x")

}
