% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pc_TRUE}
\alias{pc_TRUE}
\title{pc_TRUE}
\usage{
pc_TRUE(
  logical_vector,
  percentify = TRUE,
  NumberAndPC = FALSE,
  NArm = TRUE,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{logical_vector}{Percentage of true values in a logical vector, parsed as text (useful for reports).}

\item{percentify}{PARAM_DESCRIPTION, Default: TRUE}

\item{NumberAndPC}{PARAM_DESCRIPTION, Default: FALSE}

\item{NArm}{PARAM_DESCRIPTION, Default: TRUE}

\item{prefix}{PARAM_DESCRIPTION, Default: NULL}

\item{suffix}{PARAM_DESCRIPTION, Default: NULL}
}
\description{
FUNCTION_DESCRIPTION.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
