% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{printEveryN}
\alias{printEveryN}
\title{printEveryN}
\usage{
printEveryN(iterator = i, N = 1000)
}
\arguments{
\item{iterator}{iterator that increases (in a loop). Default: i.}

\item{N}{print Every N}
}
\description{
Report iterator value at every e.g. 1000
}
