% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{printProgress}
\alias{printProgress}
\title{Print Loop Progress}
\usage{
printProgress(i = i, total, message = "Progress", digits = 0)
}
\arguments{
\item{i}{Current iteration of the loop. Default: i.}

\item{total}{Total number of iterations in the loop.}

\item{message}{Optional message to display alongside the progress.}

\item{digits}{Digits to display}
}
\description{
Prints the progress of a loop as a number and percentage.
}
