% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pU}
\alias{pU}
\title{Print the (Number of) Unique Elements and Return an Object in a Pipe}
\usage{
pU(x, head_n = 20)
}
\arguments{
\item{x}{The object whose unique elements to print and return. Default: None.}

\item{head_n}{Max number of unique elements to print. Default: 20}
}
\value{
The input object \code{x}, unchanged.
}
\description{
Prints the unique elements of the input object and returns it, enabling you to inspect
unique values inside a pipe.
}
\examples{
results <- c(1, 2, 2, 3, 3, 3) |>
  pU() |>
  sqrt() |>
  pSee()
}
