% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{range.list}
\alias{range.list}
\title{range.list}
\usage{
\method{range}{list}(L, namesOrdered)
}
\arguments{
\item{L}{PARAM_DESCRIPTION}

\item{namesOrdered}{PARAM_DESCRIPTION}
}
\description{
Range of values in whole list.
}
