% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{range.wNames}
\alias{range.wNames}
\title{Range elements with names preserved}
\usage{
\method{range}{wNames}(x)
}
\arguments{
\item{x}{A numeric vector with names}
}
\description{
Get the range of a numeric vector, keep the names of the min and max values
}
\examples{
vec <- c(a = 1, b = 5, c = 3); range.wNames(vec)
}
