% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{remove.na.cols}
\alias{remove.na.cols}
\title{remove.na.cols}
\usage{
remove.na.cols(mat)
}
\arguments{
\item{mat}{In put matrix.}
}
\description{
Cols have to be a vector of numbers corresponding to columns.
}
