% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{remove.na.rows}
\alias{remove.na.rows}
\title{remove.na.rows}
\usage{
remove.na.rows(mat, cols = 1:NCOL(mat))
}
\arguments{
\item{mat}{In put matrix.}

\item{cols}{PARAM_DESCRIPTION, Default: 1:NCOL(mat)}
}
\description{
Cols have to be a vector of numbers corresponding to columns.
}
