% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{reorder.list}
\alias{reorder.list}
\title{reorder.list}
\usage{
\method{reorder}{list}(L, namesOrdered = gtools::mixedsort(names(L)))
}
\arguments{
\item{L}{A list.}

\item{namesOrdered}{A vector of names or indices in the desired order.}
}
\value{
A list.
}
\description{
Reorder elements of lists in your custom order of names / indices.
}
\examples{
L <- list(a = 1, b = 2, c = 3)
namesOrdered <- c("c", "a", "b")
reorder.list(L, namesOrdered)
}
\seealso{
\code{\link[gtools]{mixedsort}}
}
