% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{reverse.list.hierarchy}
\alias{reverse.list.hierarchy}
\title{reverse.list.hierarchy}
\usage{
reverse.list.hierarchy(my_list)
}
\arguments{
\item{my_list}{A list.}
}
\value{
A list with the elements of the original list in reverse order.
}
\description{
Reverses the hierarchy of a list.
}
\examples{
my_list <- list(a = list(1, 2), b = list(3, 4))
reverse.list.hierarchy(my_list)

}
