% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rotate}
\alias{rotate}
\title{rotate}
\usage{
rotate(x, clockwise = TRUE)
}
\arguments{
\item{x}{Numeric input matrix.}

\item{clockwise}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Rotate a matrix 90 degrees.
}
