% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{row.Zscore}
\alias{row.Zscore}
\title{row.Zscore}
\usage{
row.Zscore(x)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}
}
\description{
Calculate Z-score over rows of data frame.
}
