% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{row2named.vector}
\alias{row2named.vector}
\title{row2named.vector}
\usage{
row2named.vector(df_row)
}
\arguments{
\item{df_row}{data frame row}
}
\description{
Convert a dataframe row into a vector, keeping the corresponding dimension name.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
