% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowGeoMeans}
\alias{rowGeoMeans}
\title{rowGeoMeans}
\usage{
rowGeoMeans(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the median of each row of a numeric matrix / data frame.
}
