% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowMin}
\alias{rowMin}
\title{rowMin}
\usage{
rowMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input matrix, or all-numeric dataframe.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the minimum of each row of a numeric matrix / data frame.
}
\details{
#' #' @title colNrNAs
#' #' @description Calculates the number of NA-values of each column of a matrix / data frame.
#' #' @param x Numeric input matrix with the distribution.
#' #' @export
colNrNAs <- function(x) colSums(is.na(x), na.rm = FALSE)
}
