% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowMultiply}
\alias{rowMultiply}
\title{Row-wise multiplication of a matrix by a vector}
\usage{
rowMultiply(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to multiply by.}
}
\description{
Multiply each row of a matrix by the corresponding element of a vector.
}
\examples{
m <- matrix(1:8, nrow = 4, byrow = TRUE)
rowMultiply(rowDivide(m, 1:4), 1:4)
}
