% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowMutliply}
\alias{rowMutliply}
\title{rowMutliply}
\usage{
rowMutliply(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to multiply by.}
}
\description{
Mutliply by row.
}
