% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowNameMatrix}
\alias{rowNameMatrix}
\title{rowNameMatrix}
\usage{
rowNameMatrix(mat_w_dimnames)
}
\arguments{
\item{mat_w_dimnames}{A named matrix to copy from.}
}
\description{
Create a copy of your matrix, where every entry is replaced by the corresponding
row name. Useful if you want to color by row name in a plot (where you have different number of
NA-values in each row).
}
