% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowSEM}
\alias{rowSEM}
\title{rowSEM}
\usage{
rowSEM(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the SEM of each row of a numeric matrix / data frame.
}
