% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowsplit}
\alias{rowsplit}
\title{rowsplit}
\usage{
rowsplit(df, f = rownames(df))
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION, Default: rownames(df)}
}
\description{
Split a data frame by a factor corresponding to columns.
}
