% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowSubtract}
\alias{rowSubtract}
\title{rowSubtract}
\usage{
rowSubtract(mat = yy, vec = 5:1)
}
\arguments{
\item{mat}{Numeric input matrix.}

\item{vec}{Vector to subtract. Length = nr. rows.}
}
\description{
Subtract a vector (length = nr. rows) row by row from each value of the matrix
}
