% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowVariance}
\alias{rowVariance}
\title{rowVariance}
\usage{
rowVariance(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculates the CV of each ROW of a numeric matrix / data frame.
}
