% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{savehistory_Rhist}
\alias{savehistory_Rhist}
\title{Save Command History to "command_history.date.scriptname.txt"}
\usage{
savehistory_Rhist()
}
\value{
Nothing is returned, but the file path is printed to the console.
}
\description{
This function saves the command history of the current R session to a text file. The file name
includes the current date and, if available, the name of the current R script (when running in
RStudio). The file is saved in the current working directory.
}
\examples{
\dontrun{
savehistory_Rhist()
}

}
