% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{select.rows.and.columns}
\alias{select.rows.and.columns}
\title{select.rows.and.columns}
\usage{
select.rows.and.columns(df, RowIDs = NULL, ColIDs = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{RowIDs}{PARAM_DESCRIPTION, Default: NULL}

\item{ColIDs}{PARAM_DESCRIPTION, Default: NULL}
}
\description{
Subset rows and columns. It checks if the selected dimension names exist and reports if any of those they aren't found.
}
