% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{select_rows_and_columns}
\alias{select_rows_and_columns}
\title{select_rows_and_columns}
\usage{
select_rows_and_columns(df, RowIDs = NULL, ColIDs = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{RowIDs}{RowIDs, Default: NULL}

\item{ColIDs}{ColIDs, Default: NULL}
}
\description{
Subset rows and columns. It checks if the selected dimension names exist and reports if any of those they aren't found.
}
