% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sem}
\alias{sem}
\title{sem}
\usage{
sem(x, na.rm = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the standard error of the mean (SEM) for a numeric vector (it excludes NA-s by default).
}
