% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{setdiff.ls}
\alias{setdiff.ls}
\title{setdiff.ls}
\usage{
setdiff.ls(ls, ...)
}
\arguments{
\item{ls}{List of 2 or more vectors (sets) with categorical variables.}

\item{...}{Pass any other argument.}
}
\description{
Calculate set difference on a list (of 2 vectors).
}
