% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{simplify_categories}
\alias{simplify_categories}
\title{simplify_categories}
\usage{
simplify_categories(category_vec, replaceit, to)
}
\arguments{
\item{category_vec}{A vector of categories.}

\item{replaceit}{The value to replace.}

\item{to}{The replacement value.}
}
\value{
A vector with the simplified categories.
}
\description{
Replace all occurrences of \code{replaceit} in \code{category_vec} with \code{to}.
}
