% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sort.decreasing}
\alias{sort.decreasing}
\title{sort.decreasing}
\usage{
\method{sort}{decreasing}(vec)
}
\arguments{
\item{vec}{A vector.}
}
\value{
A sorted vector.
}
\description{
Sorts \code{vec} in decreasing order.
}
\examples{
sort.decreasing(rnorm(10))

}
