% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sort.mat}
\alias{sort.mat}
\title{sort.mat}
\usage{
\method{sort}{mat}(df, colname_in_df = 1, decrease = FALSE, na_last = TRUE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{colname_in_df}{PARAM_DESCRIPTION, Default: 1}

\item{decrease}{PARAM_DESCRIPTION, Default: FALSE}

\item{na_last}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Sort a matrix. ALTERNATIVE: dd\link{with(dd, order(-z, b)), }. Source: https://stackoverflow.com/questions/1296646/how-to-sort-a-dataframe-by-columns-in-r.
}
