% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sort_matrix_rows}
\alias{sort_matrix_rows}
\title{Sort matrix or data frame by a column or row names}
\usage{
sort_matrix_rows(df, column, rownames = F, decrease = FALSE, na_last = TRUE)
}
\arguments{
\item{df}{A numeric matrix or data frame to be sorted. Default: none.}

\item{column}{A column name or index by which to sort. The function can handle only single-column
sorting. Default: none.}

\item{rownames}{Logical. If \code{TRUE}, row names will be used for sorting instead of a column.
Default: \code{FALSE}.}

\item{decrease}{Logical. If \code{TRUE}, the data will be sorted in decreasing order. Default: \code{FALSE}.}

\item{na_last}{Logical. If \code{TRUE}, \code{NA} values will be placed at the end. Default: \code{TRUE}.}
}
\value{
A sorted version of the input matrix or data frame.
}
\description{
Sorts a numeric matrix or data frame by a specified column or by row names. The function can only
handle sorting by a single column. It offers options to sort in increasing or decreasing order,
and to control the placement of \code{NA} values.
}
\examples{
df <- data.frame(AA = c(1, 2, 3), BB = c(3, 2, 1), row.names = letters[7:5])
sort_matrix_rows(df, "AA")
sort_matrix_rows(df, "AA", decrease = TRUE)
sort_matrix_rows(df, "A", rownames = TRUE)

}
