% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sortbyitsnames}
\alias{sortbyitsnames}
\title{sortbyitsnames}
\usage{
sortbyitsnames(vec_or_list, decreasing = FALSE, ...)
}
\arguments{
\item{vec_or_list}{A vector or list.}

\item{decreasing}{Logical. Whether to sort in decreasing order.}

\item{...}{Additional arguments passed to \code{mixedsort()}.}
}
\value{
A sorted vector or list.
}
\description{
Sort a vector or list by the alphanumeric order of its names (instead of its values).
}
