% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{split_vec_to_list_by_N}
\alias{split_vec_to_list_by_N}
\title{Split a Vector into a List by Every N-th Element}
\usage{
split_vec_to_list_by_N(vec = 1:27, by = 9)
}
\arguments{
\item{vec}{A numeric or character vector to be split.}

\item{by}{Integer value specifying the chunk size. Default is 9.}
}
\value{
A list where each element is a vector containing up to \code{by} elements from \code{vec}.
}
\description{
This function divides a given vector into chunks of size \code{by} (default is 9).
The resulting list contains vectors of the specified chunk size or smaller.
}
