% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{splititsnames_byValues}
\alias{splititsnames_byValues}
\title{Split the names of list by its values.}
\usage{
splititsnames_byValues(namedVec)
}
\arguments{
\item{namedVec}{A vector with names.}
}
\value{
A list of vectors, each of which contains the elements of \code{namedVec} that have the corresponding value.
}
\description{
Split the names of a list by its values.
}
\examples{
namedVec <- c(a = "A", b = "B", c = "C", d = "A", e = "B", f = "D")
splititsnames_byValues(namedVec)

}
