% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sstrsplit}
\alias{sstrsplit}
\title{sstrsplit}
\usage{
sstrsplit(string, pattern = "_", n = 2)
}
\arguments{
\item{string}{PARAM_DESCRIPTION}

\item{pattern}{PARAM_DESCRIPTION, Default: '_'}

\item{n}{PARAM_DESCRIPTION, Default: 2}
}
\description{
Alias for str_split_fixed in the stringr package.
}
\seealso{
\code{\link[stringr]{str_split}}
}
