% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{stopif}
\alias{stopif}
\title{Stop Execution If Condition is True}
\usage{
stopif(condition, message = "Condition is TRUE.")
}
\arguments{
\item{condition}{A logical condition to be tested. If TRUE, an error message is thrown and execution is stopped.}

\item{message}{An optional error message to display if the condition is TRUE.}
}
\description{
This function stops the execution of the script if the provided condition evaluates to TRUE.
It is the complement of the \code{stopifnot()} function and is used for asserting conditions where
an error should be thrown if the condition is TRUE, rather than FALSE.
}
\examples{
a <- 1
stopif(a != 1, message = "A is 1")
}
