% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sumBySameName}
\alias{sumBySameName}
\title{sumBySameName}
\usage{
sumBySameName(namedVec)
}
\arguments{
\item{namedVec}{PARAM_DESCRIPTION}
}
\description{
Sum up vector elements with the same name.
}
