% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{table_decreasing}
\alias{table_decreasing}
\title{Frequency Table with Sorting Option}
\usage{
table_decreasing(vec, decreasing = TRUE, useNA = "ifany")
}
\arguments{
\item{vec}{A vector. The vector for which the frequency table is to be generated.}

\item{decreasing}{Logical. Should the table be sorted in decreasing order? Default: \code{TRUE}.}

\item{useNA}{A string. Specifies how to handle NA values. Can be "ifany", "always", or "no".
Default: \code{"ifany"}.}
}
\value{
A frequency table sorted based on the \code{decreasing} argument.
}
\description{
This function generates a frequency table of the input vector \code{vec} and allows the option
to sort the table in decreasing or increasing order. It handles NA values.
}
\examples{
vec <- c(1, 2, 2, NA, 3, 3, 3)
table_decreasing(vec)
table_decreasing(vec, decreasing = FALSE)

}
