% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{table_decreasing_w_fix}
\alias{table_decreasing_w_fix}
\title{Frequency Table with Hyrid Sorting: you can sort by frequency and by specified value}
\usage{
table_decreasing_w_fix(vec, first_elements = FALSE, useNA = "ifany")
}
\arguments{
\item{vec}{A vector. The vector for which the frequency table is to be generated.}

\item{first_elements}{Logical. Should the table be sorted in decreasing order? Default: \code{TRUE}.}

\item{useNA}{A string. Specifies how to handle NA values. Can be "ifany", "always", or "no".
Default: \code{"ifany"}.}
}
\value{
A frequency table sorted.
}
\description{
This function generates a frequency table of the input vector \code{vec} and displays the table
sorted by frequency and by a set of specified values. It handles NA values.
}
