% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{tibble_summary_to_namedVec}
\alias{tibble_summary_to_namedVec}
\title{tibble_summary_to_namedVec}
\usage{
tibble_summary_to_namedVec(
  tbl = dplyr::tibble(key = sample(x = 1:5, size = 20, replace = T), value = rnorm(20)),
  idx = c(key = 1, value = 2)
)
}
\arguments{
\item{tbl}{PARAM_DESCRIPTION, Default: dplyr::tibble(key = sample(x = 1:5, size = 20, replace = T), value = rnorm(20))}

\item{idx}{PARAM_DESCRIPTION, Default: c(key = 1, value = 2)}
}
\description{
Convert a key-value tibble into a named vector (as opposed to using rownames).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}
}
