% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{top_indices}
\alias{top_indices}
\title{top_indices}
\usage{
top_indices(x, n = 3, top = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 3}

\item{top}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Returns the position / index of the n highest values. For equal values, it maintains the original order.
}
