% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{TPM_normalize}
\alias{TPM_normalize}
\title{TPM_normalize}
\usage{
TPM_normalize(mat, SUM = 1e+06)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{SUM}{PARAM_DESCRIPTION, Default: 1e+06}
}
\description{
Normalize each column to 1 million.
}
