% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{translate}
\alias{translate}
\title{Translate a set of values to a new set using a dictionary}
\usage{
translate(vec, old, new)
}
\arguments{
\item{vec}{set of values where you want to replace}

\item{old}{oldvalues, a vector of values that occur in \code{vec} to be replaced.}

\item{new}{newvalues, a vector of equal length, to be copied from, corresponding 1-by-1 to old.}
}
\description{
Replaces a set of values in a vector with another set of values, so
it translates your vector. Oldvalues and newvalues have to be 1-to-1
corresponding vectors.  'chartr("a-cX", "D-Fw", x) does the same as above
in theory, but it did not seem very robust regarding your input...'
}
\examples{
A <- 1:3
translate(vec = A, old = 2:3, new = letters[1:2])

}
