% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{unlapply}
\alias{unlapply}
\title{Like sapply, but with names preserved}
\usage{
unlapply(list, FUN, ...)
}
\arguments{
\item{...}{Pass any other argument.}
}
\description{
Do an lapply, then unlist, finally preserve names
}
