% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{vec.from.template}
\alias{vec.from.template}
\title{vec.from.template}
\usage{
vec.from.template(x = table(LETTERS[1:5]), fill = NA)
}
\arguments{
\item{x}{A vector of names, Default: \code{LETTERS[1:5]}}

\item{fill}{The value to fill the new vector, Default: \code{NA}}
}
\description{
Create a vector from a names of another vector / list / etc.
}
