% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{vec.from.vec}
\alias{vec.from.vec}
\title{vec.from.vec}
\usage{
vec.from.vec(x = table(LETTERS[1:5]), fill = NA)
}
\arguments{
\item{x}{A vector of names, Default: LETTERS\link{1:5}}

\item{fill}{The value to fill the new vector, Default: NA}
}
\description{
Create a vector from a names of another vector / list / etc.
}
