% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{vtable}
\alias{vtable}
\title{vtable}
\usage{
vtable(x, useNA = c("no", "ifany", "always")[2], ...)
}
\arguments{
\item{x}{A vector or factor to be tabulated.}

\item{useNA}{A string specifying how to handle NA values. Can be "no", "ifany", or "always". Default: "ifany".}

\item{...}{Additional arguments passed to \code{table()}.}
}
\value{
A named vector containing the counts of each unique value in \code{x}, with names corresponding to the unique values.
}
\description{
A version of table() that returns a named vector instead of a table object.
It can handle vectors and factors, and has the same useNA argument as table().
}
