% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{what}
\alias{what}
\title{what}
\usage{
what(x, printme = 0)
}
\arguments{
\item{x}{An object to identify.}

\item{printme}{print the first "printme" elements, Default: 0}
}
\description{
A better version of is(). It can print the first "printme" elements.
}
