% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which.duplicated}
\alias{which.duplicated}
\title{which.duplicated}
\usage{
which.duplicated(vec, orig = F)
}
\arguments{
\item{vec}{input vector}

\item{orig}{PARAM_DESCRIPTION, Default: F}
}
\description{
Which values are duplicated?.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
