% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which.NA}
\alias{which.NA}
\title{which.NA}
\usage{
which.NA(vec, verbose = TRUE)
}
\arguments{
\item{vec}{input vector. Default: TRUE}

\item{verbose}{print statistics to terminal. Default: TRUE}
}
\description{
Which values are NA?.
}
