% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which_names}
\alias{which_names}
\title{which_names}
\usage{
which_names(namedVec)
}
\arguments{
\item{namedVec}{A vector of named elements.}
}
\description{
Return the names where the input vector is TRUE. The input vector is converted to logical.
}
