% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which_names_grep}
\alias{which_names_grep}
\title{which_names_grep}
\usage{
which_names_grep(namedVec, pattern, ...)
}
\arguments{
\item{namedVec}{A vector of named elements.}

\item{pattern}{A regular expression pattern.}

\item{...}{Pass any other argument to grepv()}
}
\description{
Return the vector elements whose names partially match a pattern.
}
