% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\docType{data}
\name{x}
\alias{x}
\title{Symmetric difference}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
x
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{z}{A vector.}
}
\value{
A list of vectors, each of which contains the elements that are only present in that
vector and not in any of the other vectors.
}
\description{
Quasi symmetric difference of any number of vectors.
}
\keyword{datasets}
