% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{zero.omit}
\alias{zero.omit}
\title{zero.omit}
\usage{
zero.omit(vec, verbose = TRUE)
}
\arguments{
\item{vec}{input vector}

\item{verbose}{Whether to print the range of the cleaned vector.}
}
\description{
Omit zero values from a vector.
}
