% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{any.duplicated}
\alias{any.duplicated}
\title{any.duplicated}
\usage{
\method{any}{duplicated}(vec, summarize = TRUE)
}
\arguments{
\item{vec}{input vector}

\item{summarize}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
How many entries are duplicated?.
}
