% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{any.duplicated.rownames.ls.of.df}
\alias{any.duplicated.rownames.ls.of.df}
\title{any.duplicated.rownames.ls.of.df}
\usage{
\method{any}{duplicated.rownames.ls.of.df}(ls)
}
\arguments{
\item{ls}{PARAM_DESCRIPTION}
}
\description{
Check if there are any duplocated rownames in a list of dataframes.
}
