% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.list.df.by.row}
\alias{as.list.df.by.row}
\title{as.list.df.by.row}
\usage{
\method{as.list}{df.by.row}(dtf, na.omit = TRUE, zero.omit = FALSE, omit.empty = FALSE)
}
\arguments{
\item{dtf}{PARAM_DESCRIPTION}

\item{na.omit}{PARAM_DESCRIPTION, Default: TRUE}

\item{zero.omit}{PARAM_DESCRIPTION, Default: FALSE}

\item{omit.empty}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
Split a dataframe into a list by its columns. omit.empty for the listelments; na.omit and zero.omit are applied on entries inside each list element.
}
