% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.listalike}
\alias{as.listalike}
\title{as.listalike}
\usage{
as.listalike(vec, list_wannabe)
}
\arguments{
\item{vec}{A vector.}

\item{list_wannabe}{A list with the desired dimensions.}
}
\value{
A list with the same dimensions as \code{list_wannabe}, and the same elements as \code{vec} in the same order.
}
\description{
Converts a vector to a list with the same dimensions as a given list.
}
\examples{
vec <- 1:10
list_wannabe <- list(a = 1:3, b = 4:6, c = 7:10)
as.listalike(vec, list_wannabe)

}
