% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.logical.wNames}
\alias{as.logical.wNames}
\title{as.logical.wNames}
\usage{
\method{as.logical}{wNames}(x, ...)
}
\arguments{
\item{x}{input vector}

\item{...}{Pass any other argument.}
}
\description{
Converts your input vector into a logical vector, and puts the original character values into the names of the new vector, unless it already has names.
}
