% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector.2colDF}
\alias{as.named.vector.2colDF}
\title{as.named.vector.2colDF}
\usage{
as.named.vector.2colDF(df, values = 1, names = 2, make.names = FALSE)
}
\arguments{
\item{df}{data frame}

\item{values}{Index of column with values, Default: 1}

\item{names}{Index of column with names, Default: 2}

\item{make.names}{make.names, Default: FALSE}
}
\description{
Convert a 2-column dataframe (value, name) into a named vector. Use for simple tibbles.
}
