% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as.named.vector.df}
\alias{as.named.vector.df}
\title{as.named.vector.df}
\usage{
as.named.vector.df(
  df,
  col.or.row.name.or.index = 1,
  verbose = TRUE,
  WhichDimNames = 1,
  ...
)
}
\arguments{
\item{verbose}{Print the columnname or rowname that is being used}

\item{WhichDimNames}{Shall we extract rows (2) or columns (1, default)?, Default: 1}

\item{index.rc}{Which column or row to extract (numeric index).}
}
\description{
Convert any column or row of a dataframe into a vector, keeping the corresponding dimension name.
}
\details{
#' @title sstrsplit
#'
#' @description Alias for \code{str_split_fixed} in the \code{stringr} package.
#' @param string A string.
#' @param pattern A character string pattern to split on.
#' @param n The number of elements to return in each split.
#' @return A list of character vectors.
#' @seealso
#'  \code{\link[stringr]{str_split}}
#' @importFrom stringr str_split_fixed
#' @export
sstrsplit <- function(string, pattern = "_", n = F) {
if (!n) n <- stringr::str_count(string = string, pattern = pattern)
stringr::str_split_fixed(string, pattern = pattern, n = n)
}
}
