% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{as_tibble_from_namedVec}
\alias{as_tibble_from_namedVec}
\title{as_tibble_from_namedVec}
\usage{
as_tibble_from_namedVec(vec.w.names = c(a = 1, b = 2), transpose = TRUE)
}
\arguments{
\item{vec.w.names}{A vector with names, Default: c(a = 1, b = 2)}

\item{transpose}{Transpose? Default: TRUE}
}
\description{
Convert a vector with names into a tibble, keeping the names as rownames.
}
\examples{
as_tibble_from_namedVec()
}
