% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{clip.at.fixed.value}
\alias{clip.at.fixed.value}
\title{clip.at.fixed.value}
\usage{
clip.at.fixed.value(x, high = TRUE, thr = 3)
}
\arguments{
\item{x}{A vector of numeric values (distribution).}

\item{high}{Clip above threshold? Default: TRUE}

\item{thr}{threshold values, Default: 3}
}
\description{
Signal clipping. Cut values in a distribution, above or below a threshold.
}
