% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{clip.at.fixed.value}
\alias{clip.at.fixed.value}
\title{clip.at.fixed.value}
\usage{
clip.at.fixed.value(x, thr, above = TRUE)
}
\arguments{
\item{x}{A vector of numeric values (distribution).}

\item{thr}{threshold values.}

\item{above}{Clip above threshold, or below? Default: TRUE (= low pass filter).}
}
\description{
Signal clipping. Cut values in a distribution, above or below a threshold.
}
