% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{clip.outliers.at.percentile}
\alias{clip.outliers.at.percentile}
\title{clip.outliers.at.percentile}
\usage{
clip.outliers.at.percentile(
  distribution,
  high = TRUE,
  percentiles = c(0.01, 0.99),
  na.rm = TRUE,
  showhist = FALSE,
  ...
)
}
\arguments{
\item{distribution}{A vector of numeric values.}

\item{high}{Clip above threshold? Default: TRUE}

\item{percentiles}{At which percentiles to cut off?, Default: c(0.01, 0.99)}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}

\item{showhist}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{Pass any other argument.}
}
\description{
Signal clipping based on the input data's distribution. It clips values above or
below the extreme N\% of the distribution.
}
