% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colDivide}
\alias{colDivide}
\title{colDivide}
\usage{
colDivide(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to divide by.}
}
\description{
Divide by column. See more: https://stackoverflow.com/questions/20596433/how-to-divide-each-row-of-a-matrix-by-elements-of-a-vector-in-r.
}
