% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colMultiply}
\alias{colMultiply}
\title{Column-wise multiplication of a matrix by a vector}
\usage{
colMultiply(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to multiply by.}
}
\description{
Multiply each column of a matrix by the corresponding element of a vector. See more: https://stackoverflow.com/questions/20596433/how-to-divide-each-row-of-a-matrix-by-elements-of-a-vector-in-r.
}
\examples{
m <- matrix(1:8, nrow = 4, byrow = TRUE)
colMultiply(colDivide(m), colSums(m))
}
