% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{colQuantile}
\alias{colQuantile}
\title{colQuantile}
\usage{
colQuantile(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}

\item{...}{Pass any other argument.}
}
\description{
Calculates the quantile of each column of a numeric matrix / data frame.
}
