% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{copy.dimension.and.dimnames}
\alias{copy.dimension.and.dimnames}
\title{copy.dimension.and.dimnames}
\usage{
copy.dimension.and.dimnames(list.1D, obj.2D)
}
\arguments{
\item{list.1D}{PARAM_DESCRIPTION}

\item{obj.2D}{PARAM_DESCRIPTION}
}
\description{
Copy dimension and dimnames.
}
