% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{cumsubtract}
\alias{cumsubtract}
\title{cumsubtract}
\usage{
cumsubtract(numericVec)
}
\arguments{
\item{numericVec}{Numeric vector to compute pairwise differences.}
}
\description{
Cumulative subtraction, opposite of cumsum().
}
