% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{df.col.2.named.vector}
\alias{df.col.2.named.vector}
\title{df.col.2.named.vector}
\usage{
df.col.2.named.vector(df_col)
}
\arguments{
\item{df_col}{data frame column}
}
\description{
Convert a dataframe column into a vector, keeping the corresponding dimension name.
}
