% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{df.remove.empty.rows.and.columns}
\alias{df.remove.empty.rows.and.columns}
\title{df.remove.empty.rows.and.columns}
\usage{
df.remove.empty.rows.and.columns(
  df = UVI.assignment.filtered.3.HF,
  suffix = substitute(df),
  rows = "rows",
  cols = "cols",
  empty = 0,
  plot_stats = T,
  ...
)
}
\arguments{
\item{df}{Data frame to filter for rows and columns with only-0 values.}

\item{cols}{}

\item{empty}{what is empty? Default: 0}

\item{plot_stats}{plot removal stats?}
}
\description{
df.remove.empty.rows.and.columns
}
