% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{df.row.2.named.vector}
\alias{df.row.2.named.vector}
\title{df.row.2.named.vector}
\usage{
df.row.2.named.vector(df_row)
}
\arguments{
\item{df_row}{data frame row}
}
\description{
Convert a dataframe row into a vector, keeping the corresponding dimension name.
}
