% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{geomean}
\alias{geomean}
\title{geomean}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Should NA values be removed before calculation? Defaults to TRUE.}
}
\value{
The geometric mean of the input vector.
}
\description{
Calculates the geometric mean of a numeric vector (it excludes NA-s by default).
}
\examples{
geomean(1:5)

}
