% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{get.oddoreven}
\alias{get.oddoreven}
\title{get.oddoreven}
\usage{
get.oddoreven(df_ = NULL, rows = FALSE, odd = TRUE)
}
\arguments{
\item{df_}{A data frame. Default: NULL}

\item{rows}{Logical indicating whether to return rows (if TRUE) or columns (if FALSE). Default: FALSE}

\item{odd}{Logical indicating whether to return odd indexed rows/columns (if TRUE) or even indexed rows/columns (if FALSE). Default: TRUE}
}
\value{
A subset of the input data frame.
}
\description{
Returns either odd or even indexed rows or columns from a data frame.
}
