% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{getCols}
\alias{getCols}
\title{getCols}
\usage{
getCols(
  mat,
  colnamez,
  silent = FALSE,
  removeNAonly = FALSE,
  remove0only = FALSE
)
}
\arguments{
\item{mat}{In put matrix.}

\item{colnamez}{PARAM_DESCRIPTION}

\item{silent}{PARAM_DESCRIPTION, Default: FALSE}

\item{removeNAonly}{PARAM_DESCRIPTION, Default: FALSE}

\item{remove0only}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
Get the subset of cols with existing colnames, report how much it could not find.
}
