% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{getRows}
\alias{getRows}
\title{getRows}
\usage{
getRows(
  mat,
  rownamez,
  silent = FALSE,
  removeNAonly = FALSE,
  remove0only = FALSE
)
}
\arguments{
\item{mat}{In put matrix.}

\item{rownamez}{PARAM_DESCRIPTION}

\item{silent}{PARAM_DESCRIPTION, Default: FALSE}

\item{removeNAonly}{PARAM_DESCRIPTION, Default: FALSE}

\item{remove0only}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
Get the subset of rows with existing rownames, report how much it could not find.
}
