% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{idimnames}
\alias{idimnames}
\title{idimnames}
\usage{
idimnames(any_object)
}
\arguments{
\item{any_object}{PARAM_DESCRIPTION}
}
\description{
A dimnames() function that can handle if you pass on a vector: it gives back the names.
}
