% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{imovingSEM}
\alias{imovingSEM}
\title{imovingSEM}
\usage{
imovingSEM(x, oneSide = 5)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{oneSide}{PARAM_DESCRIPTION, Default: 5}
}
\description{
Calculates the moving / rolling standard error of the mean (SEM). It calculates it to the edge of the vector with incrementally smaller window-size.
}
