% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{imovingSEM}
\alias{imovingSEM}
\title{imovingSEM}
\usage{
imovingSEM(x, oneSide = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{oneSide}{The size of the moving window, in terms of the number of elements on either side of the current element.}
}
\value{
A vector of the same length as \code{x}, containing the SEMs for each element.
}
\description{
Calculates the moving / rolling standard error of the mean (SEM). It calculates it to the edge of the vector with incrementally smaller window-size.
}
