% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intermingle.cbind}
\alias{intermingle.cbind}
\title{intermingle.cbind}
\usage{
intermingle.cbind(df1, df2)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame.}
}
\value{
A data frame that combines \code{df1} and \code{df2}, with the columns of \code{df1} alternating with the columns of `
}
\description{
Combine 2 data frames (of the same length) so that form every odd and every even element of a unified list. Useful for side-by-side comparisons, e.g. in wstripchart_list().
}
