% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intermingle.cbind}
\alias{intermingle.cbind}
\title{intermingle.cbind}
\usage{
intermingle.cbind(df1, df2)
}
\arguments{
\item{df1}{PARAM_DESCRIPTION}

\item{df2}{PARAM_DESCRIPTION}
}
\description{
Combine 2 data frames (of the same length) so that form every odd and every even element of a unified list. Useful for side-by-side comparisons, e.g. in wstripchart_list().
}
