% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{intersect.ls}
\alias{intersect.ls}
\title{intersect.ls}
\usage{
intersect.ls(ls, ...)
}
\arguments{
\item{ls}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
Intersect any number of lists.
}
