% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.2.replicated.name.vec}
\alias{list.2.replicated.name.vec}
\title{list.2.replicated.name.vec}
\usage{
list.2.replicated.name.vec(ListWithNames = Sections.ls.Final)
}
\arguments{
\item{ListWithNames}{A list.}
}
\value{
A vector with the same elements as \code{ListWithNames}, but with the names of the list elements repeated as many times as the number of elements in each list element.
}
\description{
Converts a list to a vector, with list elements names replicated as many times as many elements each element had.
}
\examples{
ListWithNames <- list(a = 1, b = 2, c = 3)
list.2.replicated.name.vec(ListWithNames)
}
