% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list.2.replicated.name.vec}
\alias{list.2.replicated.name.vec}
\title{list.2.replicated.name.vec}
\usage{
list.2.replicated.name.vec(ListWithNames = Sections.ls.Final)
}
\arguments{
\item{ListWithNames}{PARAM_DESCRIPTION, Default: Sections.ls.Final}
}
\description{
Convert a list to a vector, with list elements names replicated as many times, as many elements each element had.
}
