% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list2fullDF.byNames}
\alias{list2fullDF.byNames}
\title{list2fullDF.byNames}
\usage{
list2fullDF.byNames(
  your.list = list(set.1 = vec.fromNames(LETTERS[1:5], fill = 1), set.2 =
    vec.fromNames(LETTERS[3:9], fill = 2)),
  byRow = TRUE,
  FILL = NA
)
}
\arguments{
\item{your.list}{PARAM_DESCRIPTION, Default: list(set.1 = vec.fromNames(LETTERS\link{1:5}, fill = 1), set.2 = vec.fromNames(LETTERS\link{3:9},
fill = 2))}

\item{byRow}{PARAM_DESCRIPTION, Default: TRUE}

\item{FILL}{PARAM_DESCRIPTION, Default: NA}
}
\description{
..
}
