% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list2fullDF.byNames}
\alias{list2fullDF.byNames}
\title{list2fullDF.byNames}
\usage{
list2fullDF.byNames(
  your.list = list(set.1 = vec.fromNames(LETTERS[1:5], fill = 1), set.2 =
    vec.fromNames(LETTERS[3:9], fill = 2)),
  as.df = TRUE,
  byRow = TRUE,
  FILL = NA
)
}
\arguments{
\item{your.list}{A list.}

\item{byRow}{Logical. Whether the resulting matrix should be arranged by row (default) or by column.}

\item{FILL}{A value to fill in missing entries.}
}
\value{
A matrix with the same elements as \code{your.list}, but with rows and columns named by the elements of the list.
}
\description{
Converts a list to a full matrix, with rows and columns named by the elements of the list.
}
\examples{
your.list <- list(set.1 = LETTERS[1:5], set.2 = LETTERS[3:9])
list2fullDF.byNames(your.list)

}
