% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{list2fullDF.presence}
\alias{list2fullDF.presence}
\title{list2fullDF.presence}
\usage{
list2fullDF.presence(
  your.list = list(set.1 = LETTERS[1:5], set.2 = LETTERS[3:9]),
  byRow = TRUE,
  FILL = 0
)
}
\arguments{
\item{your.list}{A list.}

\item{byRow}{Logical. Whether the resulting matrix should be arranged by row (default) or by column.}

\item{FILL}{A value to fill in missing entries.}
}
\value{
A matrix with the same elements as \code{your.list}, but with rows and columns named by the elements of the list.
}
\description{
Converts a list to a full matrix, with rows and columns named by the elements of the list.

The matrix will contain a 1 in each cell where the corresponding element of the list is present, and a 0 otherwise.
}
\examples{
your.list <- list("set.1" = LETTERS[1:5], "set.2" = LETTERS[3:9])
list2fullDF.presence(your.list)

}
