% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{ls2categvec}
\alias{ls2categvec}
\title{ls2categvec}
\usage{
ls2categvec(your_list)
}
\arguments{
\item{your_list}{PARAM_DESCRIPTION}
}
\description{
Convert a list to a vector repeating list-element names, while vector names are the list elements.
}
