% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{matrix.fromVector}
\alias{matrix.fromVector}
\title{matrix.fromVector}
\usage{
matrix.fromVector(vector = 1:5, HowManyTimes = 3, IsItARow = TRUE)
}
\arguments{
\item{vector}{Input vector for rows (number of rows = length), Default: 1:5}

\item{HowManyTimes}{Number of columns, Default: 3}

\item{IsItARow}{Transpose? Swap rows an columns. Default: TRUE}
}
\description{
Create a matrix from values in a vector repeated for each column / each row. Similar to rowNameMatrix and colNameMatrix.
}
