% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{mdlapply}
\alias{mdlapply}
\title{mdlapply}
\usage{
mdlapply(list_2D, ...)
}
\arguments{
\item{list_2D}{PARAM_DESCRIPTION}

\item{...}{Pass any other argument.}
}
\description{
lapply for multidimensional arrays.
}
