% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_1col_dfs_by_rn}
\alias{merge_1col_dfs_by_rn}
\title{merge_1col_dfs_by_rn}
\usage{
merge_1col_dfs_by_rn(list_of_dfs, FILLwith = 0, columnUSE = "n")
}
\arguments{
\item{list_of_dfs}{list of 1col dfs}

\item{FILLwith}{0 by def}

\item{columnUSE}{colnames in both}
}
\description{
merge_1col_dfs_by_rn
}
\examples{
merge_1col_dfs_by_rn()
}
