% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{merge_dfs_by_rn}
\alias{merge_dfs_by_rn}
\title{merge_dfs_by_rn}
\usage{
merge_dfs_by_rn(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{PARAM_DESCRIPTION}
}
\description{
Merge any data frames by rownames. Required plyr package.
}
\seealso{
\code{\link[plyr]{join_all}}
}
