% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{movingAve}
\alias{movingAve}
\title{movingAve}
\usage{
movingAve(x, oneSide = 5)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{oneSide}{PARAM_DESCRIPTION, Default: 5}
}
\description{
Calculates the moving / rolling average of a numeric vector.
}
