% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{movingAve2}
\alias{movingAve2}
\title{Moving / rolling average (v2, filter)}
\usage{
movingAve2(x, n = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The size of the moving window. Defaults to 5.}
}
\value{
A vector of the moving averages.
}
\description{
Calculates the moving / rolling average of a numeric vector, using \code{filter()}.
}
\examples{
movingAve2(1:5)

}
