% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{movingAve2}
\alias{movingAve2}
\title{movingAve2}
\usage{
movingAve2(x, n = 5)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 5}
}
\description{
Calculates the moving / rolling average of a numeric vector, using filter().
}
