% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{movingSEM}
\alias{movingSEM}
\title{movingSEM}
\usage{
movingSEM(x, oneSide = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{oneSide}{The size of the moving window, in terms of the number of elements on either side of the current element.}
}
\value{
A vector of the same length as \code{x}, containing the SEMs for each element.
}
\description{
Calculates the moving / rolling standard error of the mean (SEM) on a numeric vector.
}
