% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{na.omit.mat}
\alias{na.omit.mat}
\title{na.omit.mat}
\usage{
\method{na.omit}{mat}(mat, any = TRUE)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{any}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Omit rows with NA values from a matrix. Rows with any, or full of NA-s.
}
