% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{numerate}
\alias{numerate}
\title{Formats a Sequence of Numbers with Zero Padding}
\usage{
numerate(
  x = 1,
  y = 100,
  zeropadding = TRUE,
  pad_length = floor(log10(max(abs(x), abs(y)))) + 1
)
}
\arguments{
\item{x}{The starting number of the sequence. Default: 1.}

\item{y}{The ending number of the sequence. Default: 100.}

\item{zeropadding}{Logical; whether to pad numbers with zeros. Default: TRUE.}

\item{pad_length}{The length of padding for the numbers. Default: Calculated as
floor(log10(max(abs(x), abs(y)))) + 1.}
}
\value{
A vector of formatted numbers, with or without zero padding.
}
\description{
This function generates a sequence of numbers between two specified values,
optionally padding them with leading zeros to a specified length. It is useful
for creating numeric sequences with consistent character lengths.
}
\examples{
numerate(1, 122)
}
\seealso{
\code{\link[stringr]{str_pad}}
}
