% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{numerate}
\alias{numerate}
\title{numerate}
\usage{
numerate(
  x = 1,
  y = 100,
  zeropadding = TRUE,
  pad_length = floor(log10(max(abs(x), abs(y)))) + 1
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: 1}

\item{y}{PARAM_DESCRIPTION, Default: 100}

\item{zeropadding}{PARAM_DESCRIPTION, Default: TRUE}

\item{pad_length}{PARAM_DESCRIPTION, Default: floor(log10(max(abs(x), abs(y)))) + 1}
}
\description{
Numerate from x to y with additonal zeropadding.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_pad}}
}
