% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pad.na}
\alias{pad.na}
\title{pad.na}
\usage{
pad.na(x, len)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{len}{PARAM_DESCRIPTION}
}
\description{
Fill up with a vector to a given length with NA-values at the end.
}
