% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pc_in_total_of_match}
\alias{pc_in_total_of_match}
\title{pc_in_total_of_match}
\usage{
pc_in_total_of_match(vec_or_table, category, NA_omit = TRUE)
}
\arguments{
\item{vec_or_table}{PARAM_DESCRIPTION}

\item{category}{PARAM_DESCRIPTION}

\item{NA_omit}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Percentage of a certain value within a vector or table.
}
\seealso{
\code{\link[stats]{na.omit}}
}
