% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pLength}
\alias{pLength}
\title{Print Length and Return an Object in a Pipe}
\usage{
pLength(x)
}
\arguments{
\item{x}{The object whose length to print and return. Default: None.}
}
\value{
The input object \code{x}, unchanged.
}
\description{
Prints the length of the input object and returns it, allowing you to verify
length inside a pipe operation.
}
\examples{
results <- c(9:1) |>
  tail(4)  |>
  pLength() |>
  sqrt()
results
}
