% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{printEveryN}
\alias{printEveryN}
\title{printEveryN}
\usage{
printEveryN(i = i, N = 1000)
}
\arguments{
\item{N}{print Every N}

\item{iterator}{Current iteration of the loop. Default: i.}
}
\description{
Report iterator value at every e.g. 1000
}
