% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{pSee}
\alias{pSee}
\title{Print and Return an Object in a Pipe}
\usage{
pSee(x, max_elements = 100)
}
\arguments{
\item{x}{The object to print and return. Default: None.}

\item{max_elements}{Number of elements of \code{x} to print. Default: 100.}
}
\value{
The input object \code{x}, unchanged.
}
\description{
Prints the input object and returns it, enabling you to inspect values inside a pipe.
}
\examples{
results <- c(1, 2, 3) \%>\%
  pSee() \%>\%
  sqrt() \%>\%
  tail(2)
results
}
