% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{range.list}
\alias{range.list}
\title{range.list}
\usage{
\method{range}{list}(L, namesOrdered)
}
\arguments{
\item{L}{A list.}
}
\value{
A vector of length 2, containing the minimum and maximum values in \code{L}.
}
\description{
Calculates the range of values in a list.
}
