% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers}
\usage{
remove_outliers(x, na.rm = TRUE, probs = c(0.05, 0.95), ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE.}

\item{probs}{A vector of two probabilities, specifying the trailing percentiles to use.}

\item{...}{Additional arguments passed to \code{quantile()}.}
}
\value{
A vector with the outliers removed.
}
\description{
Remove values that fall outside the trailing \code{probs} percentiles of the distribution.
}
