% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers}
\usage{
remove_outliers(x, na.rm = TRUE, ..., probs = c(0.05, 0.95))
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}

\item{probs}{PARAM_DESCRIPTION, Default: c(0.05, 0.95)}
}
\description{
Remove values that fall outside the trailing N \% of the distribution.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
