% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{reorder.list}
\alias{reorder.list}
\title{reorder.list}
\usage{
\method{reorder}{list}(L, namesOrdered = mixedsort(names(L)))
}
\arguments{
\item{L}{PARAM_DESCRIPTION}

\item{namesOrdered}{PARAM_DESCRIPTION, Default: mixedsort(names(L))}
}
\description{
Reorder elements of lists in your custom order of names / indices.
}
