% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rescale}
\alias{rescale}
\title{rescale}
\usage{
rescale(vec, from = 0, upto = 100)
}
\arguments{
\item{vec}{input vector}

\item{from}{PARAM_DESCRIPTION, Default: 0}

\item{upto}{PARAM_DESCRIPTION, Default: 100}
}
\description{
Linear transformation to a given range of values.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
