% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{reverse.list.hierarchy}
\alias{reverse.list.hierarchy}
\title{reverse.list.hierarchy}
\usage{
reverse.list.hierarchy(ll)
}
\arguments{
\item{ll}{PARAM_DESCRIPTION}
}
\description{
Reverse list hierarchy.
}
