% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rotate_matrix}
\alias{rotate_matrix}
\title{Rotate a Matrix by 90 Degrees}
\usage{
rotate_matrix(x, clockwise = TRUE)
}
\arguments{
\item{x}{A numeric matrix that is to be rotated.}

\item{clockwise}{Logical; if TRUE (default), rotates the matrix 90 degrees clockwise.}
}
\value{
A numeric matrix rotated 90 degrees in the specified direction.
}
\description{
Rotates a given numeric matrix 90 degrees in a specified direction. The rotation
can be either clockwise or counterclockwise, determined by the \code{clockwise} parameter.
}
\examples{
# Define a 3x3 matrix
matrix_original <- matrix(1:9, nrow = 3)

# Rotate the matrix clockwise
rotated_clockwise <- rotate(matrix_original, TRUE)

# Rotate the matrix counterclockwise
rotated_counterclockwise <- rotate(matrix_original, FALSE)

}
