% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowDivide}
\alias{rowDivide}
\title{rowDivide}
\usage{
rowDivide(mat, vec = rowSums(mat))
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to divide by. Default: row sums of the matrix.}
}
\description{
Divide by row.
}
\examples{
rowDivide(rowMultiply(m, 1:3), 1:3)
}
