% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowDivide}
\alias{rowDivide}
\title{rowDivide}
\usage{
rowDivide(mat, vec)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{vec}{Vector to divide by.}
}
\description{
Divide by row.
}
