% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{rowMax}
\alias{rowMax}
\title{rowMax}
\usage{
rowMax(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric input matrix with the distribution.}

\item{na.rm}{Remove NA values for calculation? Default: TRUE}
}
\description{
Calculates the maximum of each row of a numeric matrix / data frame.
}
