% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{setIfNotDefined}
\alias{setIfNotDefined}
\title{Set object if not defined}
\usage{
setIfNotDefined(obj, value_expr)
}
\arguments{
\item{obj}{Unquoted symbol to check / define.}

\item{value_expr}{Expression evaluated only if \code{obj} is missing.}
}
\value{
The existing or newly created object.
}
\description{
Checks whether a symbol exists in the calling environment. If it does not,
the expression provided as the second argument is evaluated, assigned to
that symbol, and returned to the global env. If it already exists, the existing object is
returned unchanged.
}
\examples{
rm(list = "x", inherits = FALSE)
setIfNotDefined(x, 22)
x <- 33
setIfNotDefined(x, { stop("must not run") }); x

}
