% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{simplify_categories}
\alias{simplify_categories}
\title{simplify_categories}
\usage{
simplify_categories(category_vec, replaceit, to)
}
\arguments{
\item{category_vec}{PARAM_DESCRIPTION}

\item{replaceit}{PARAM_DESCRIPTION}

\item{to}{PARAM_DESCRIPTION}
}
\description{
Replace every entry that is found in "replaceit", by a single value provided by "to".
}
