% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sort.decreasing}
\alias{sort.decreasing}
\title{sort.decreasing}
\usage{
\method{sort}{decreasing}(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
Sort in decreasing order.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
