% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sortbyitsnames}
\alias{sortbyitsnames}
\title{sortbyitsnames}
\usage{
sortbyitsnames(vec_or_list, decreasing = FALSE, ...)
}
\arguments{
\item{vec_or_list}{PARAM_DESCRIPTION}

\item{decreasing}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{Pass any other argument.}
}
\description{
Sort a vector by the alphanumeric order of its names(instead of its values).
}
\seealso{
\code{\link[gtools]{mixedsort}}
}
