% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sortEachColumn}
\alias{sortEachColumn}
\title{sortEachColumn}
\usage{
sortEachColumn(data, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{...}{Pass any other argument.}
}
\description{
Sort each column of a numeric matrix / data frame.
}
