% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{splititsnames_byValues}
\alias{splititsnames_byValues}
\title{splititsnames_byValues}
\usage{
splititsnames_byValues(namedVec)
}
\arguments{
\item{namedVec}{Vector with names}
}
\description{
Split a list by its names.
}
