% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{sstrsplit}
\alias{sstrsplit}
\title{sstrsplit}
\usage{
sstrsplit(string, pattern = "_", n = 2)
}
\arguments{
\item{string}{A string.}

\item{pattern}{A character string pattern to split on.}

\item{n}{The number of elements to return in each split.}
}
\value{
A list of character vectors.
}
\description{
Alias for \code{str_split_fixed} in the \code{stringr} package.
}
\seealso{
\code{\link[stringr]{str_split}}
}
