% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{symdiff}
\alias{symdiff}
\title{symdiff}
\usage{
symdiff(x, y, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{...}{Pass any other argument.}
}
\description{
Quasy symmetric difference of any number of vectors.
}
