% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{table_fixed_categories}
\alias{table_fixed_categories}
\title{table_fixed_categories}
\usage{
table_fixed_categories(vec, categories_vec)
}
\arguments{
\item{vec}{Input vector}

\item{categories_vec}{PARAM_DESCRIPTION}
}
\description{
Generate a table() with a fixed set of categories. It fills up the table with missing categories, that are relevant when comparing to other vectors.
}
\examples{
\dontrun{
if(interactive()){
 # table_fixed_categories(vec = 1:3, categories_vec = 1:4)
 }
}
}
