% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{top_indices}
\alias{top_indices}
\title{top_indices}
\usage{
top_indices(x, n = 3, top = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{n}{The number of highest values to return.}

\item{top}{Whether to return the \code{n} highest (or \code{n} lowest) values.}
}
\value{
A vector of indices.
}
\description{
Returns the positions of the \code{n} highest values in \code{x}.
}
\examples{
top_indices(rnorm(10), n = 3)

}
