% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{TPM_normalize}
\alias{TPM_normalize}
\title{TPM_normalize}
\usage{
TPM_normalize(mat, SUM = 1e+06)
}
\arguments{
\item{mat}{Numeric input matrix with the distribution.}

\item{SUM}{PARAM_DESCRIPTION, Default: 1e+06}
}
\description{
Normalize each column to 1 million.
}
\details{
#' #' @title colNrNAs
#' #' @description Calculates the number of NA-values of each column of a matrix / data frame.
#' #' @param x Numeric input matrix with the distribution.
#' #' @export
colNrNAs <- function(x) colSums(is.na(x), na.rm = FALSE)
}
