% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{trail}
\alias{trail}
\title{trail}
\usage{
trail(vec, N = 10)
}
\arguments{
\item{vec}{input vector}

\item{N}{PARAM_DESCRIPTION, Default: 10}
}
\description{
A combination of head() and tail() to see both ends.
}
