% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{trail}
\alias{trail}
\title{trail}
\usage{
trail(vec, N = 10)
}
\arguments{
\item{vec}{A vector.}

\item{N}{The number of elements to include from each end.}
}
\value{
A vector.
}
\description{
A combination of \code{head()} and \code{tail()}: Returns a vector containing the first
\code{N} and last \code{N} elements of vector.
}
\examples{
trail(rnorm(10), N = 3)

}
