% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{unique.wNames}
\alias{unique.wNames}
\title{Unique elements}
\usage{
\method{unique}{wNames}(x)
}
\arguments{
\item{x}{A vector with names}
}
\description{
Get the unique elements of a vector, keep their names
}
