% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{vec.fromNames}
\alias{vec.fromNames}
\title{vec.fromNames}
\usage{
vec.fromNames(name_vec = LETTERS[1:5], fill = NA)
}
\arguments{
\item{name_vec}{A vector of names, Default: \code{LETTERS[1:5]}}

\item{fill}{The value to fill the new vector, Default: \code{NA}}
}
\description{
Create a vector from a vector of names.
}
