% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{what}
\alias{what}
\title{what}
\usage{
what(x, printme = 0)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{printme}{PARAM_DESCRIPTION, Default: 0}
}
\description{
A better version of is(). It can print the first "printme" elements.
}
\examples{
\dontrun{
if(interactive()){
 # a <- 1:2; what(a)
 }
}
}
