% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which.duplicated}
\alias{which.duplicated}
\title{which.duplicated}
\usage{
which.duplicated(vec, verbose = TRUE)
}
\arguments{
\item{vec}{input vector}

\item{verbose}{print statistics to terminal. Default: TRUE}
}
\description{
Which values are duplicated?.
}
