% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeAndRoll2.R
\name{which_names_grep}
\alias{which_names_grep}
\title{which_names_grep}
\usage{
which_names_grep(namedVec, pattern)
}
\arguments{
\item{namedVec}{PARAM_DESCRIPTION}

\item{pattern}{PARAM_DESCRIPTION}
}
\description{
Return the vector elements whose names are partially matched.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
